app/prompt.py
from langchain_core.prompts import ChatPromptTemplate

template = """
Answer given the following context:
{context}

Question: {question}
"""

ANSWER_PROMPT = ChatPromptTemplate.from_template(template)